Feature #70119, refactor VSInputService as service proxy, update related controllers
