+++ core/src/main/java/inetsoft/web/composer/model/TreeNodeModel.java
@@ -20,12 +20,14 @@ package inetsoft.web.composer.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import org.immutables.serial.Serial;
 import org.immutables.value.Value;
 
 import javax.annotation.Nullable;
 import java.util.*;
 
 @Value.Immutable
+@Serial.Structural
 @JsonSerialize(as = ImmutableTreeNodeModel.class)
 @JsonDeserialize(as = ImmutableTreeNodeModel.class)
 @JsonInclude(JsonInclude.Include.NON_NULL)
