+++ core/src/main/java/inetsoft/web/composer/model/vs/BasicGeneralPaneModel.java
@@ -19,8 +19,10 @@ package inetsoft.web.composer.model.vs;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
+import java.io.Serializable;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class BasicGeneralPaneModel {
+public class BasicGeneralPaneModel implements Serializable {
    public String getName() {
       return name;
    }
