+++ core/src/main/java/inetsoft/web/composer/model/vs/ClickableScriptPaneModel.java
@@ -20,6 +20,7 @@ package inetsoft.web.composer.model.vs;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import org.immutables.serial.Serial;
 import org.immutables.value.Value;
 
 /**
@@ -27,6 +28,7 @@ import org.immutables.value.Value;
  * oval property dialog
  */
 @Value.Immutable
+@Serial.Structural
 @JsonSerialize(as = ImmutableClickableScriptPaneModel.class)
 @JsonDeserialize(as = ImmutableClickableScriptPaneModel.class)
 @JsonIgnoreProperties(ignoreUnknown = true)
