+++ core/src/main/java/inetsoft/web/composer/model/vs/ColumnOptionDialogModel.java
@@ -19,8 +19,10 @@ package inetsoft.web.composer.model.vs;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
+import java.io.Serializable;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ColumnOptionDialogModel {
+public class ColumnOptionDialogModel implements Serializable {
 
    public boolean isEnableColumnEditing() {
       return enableColumnEditing;
