+++ core/src/main/java/inetsoft/web/composer/model/vs/ComboBoxDefaultValueListModel.java
@@ -17,7 +17,9 @@
  */
 package inetsoft.web.composer.model.vs;
 
-public class ComboBoxDefaultValueListModel {
+import java.io.Serializable;
+
+public class ComboBoxDefaultValueListModel implements Serializable {
 
    public String getLabel() {
       return label;
