+++ core/src/main/java/inetsoft/web/composer/model/vs/ListValuesPaneModel.java
@@ -19,8 +19,10 @@ package inetsoft.web.composer.model.vs;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
+import java.io.Serializable;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ListValuesPaneModel {
+public class ListValuesPaneModel implements Serializable {
    public int getSortType() {
       return sortType;
    }
