+++ core/src/main/java/inetsoft/web/composer/model/vs/NumericRangePaneModel.java
@@ -17,7 +17,9 @@
  */
 package inetsoft.web.composer.model.vs;
 
-public class NumericRangePaneModel {
+import java.io.Serializable;
+
+public class NumericRangePaneModel implements Serializable {
    public String getMaximum() {
       return maximum;
    }
