+++ core/src/main/java/inetsoft/web/composer/vs/dialog/VSSortingDialogController.java
@@ -17,20 +17,6 @@
  */
 package inetsoft.web.composer.vs.dialog;
 
-
-import inetsoft.analytic.composition.ViewsheetService;
-import inetsoft.report.StyleConstants;
-import inetsoft.report.composition.RuntimeViewsheet;
-import inetsoft.uql.ColumnSelection;
-import inetsoft.uql.asset.*;
-import inetsoft.uql.erm.DataRef;
-import inetsoft.uql.viewsheet.CalcTableVSAssembly;
-import inetsoft.uql.viewsheet.Viewsheet;
-import inetsoft.uql.viewsheet.internal.CalcTableVSAssemblyInfo;
-import inetsoft.util.Tool;
-import inetsoft.web.binding.drm.DataRefModel;
-import inetsoft.web.binding.handler.VSAssemblyInfoHandler;
-import inetsoft.web.binding.service.DataRefModelFactoryService;
 import inetsoft.web.composer.model.vs.*;
 import inetsoft.web.viewsheet.Undoable;
 import inetsoft.web.viewsheet.model.RuntimeViewsheetRef;
@@ -41,7 +27,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import java.security.Principal;
-import java.util.*;
 
 /**
  * Controller that provides the endpoints for the vs sorting dialog.
@@ -54,17 +39,11 @@ public class VSSortingDialogController {
     * Creates a new instance of <tt>VSSortingDialogController</tt>.
     */
    @Autowired
-   public VSSortingDialogController(
-      VSObjectService service, VSInputService vsInputService,
-      DataRefModelFactoryService dataRefModelFactoryService, ViewsheetService viewsheetService,
-      RuntimeViewsheetRef runtimeViewsheetRef, VSAssemblyInfoHandler vsAssemblyInfoHandler)
+   public VSSortingDialogController(VSInputServiceProxy vsInputServiceProxy,
+                                    RuntimeViewsheetRef runtimeViewsheetRef)
    {
-      this.service = service;
-      this.vsInputService = vsInputService;
-      this.dataRefModelFactoryService = dataRefModelFactoryService;
-      this.viewsheetService = viewsheetService;
+      this.vsInputServiceProxy = vsInputServiceProxy;
       this.runtimeViewsheetRef = runtimeViewsheetRef;
-      this.vsAssemblyInfoHandler = vsAssemblyInfoHandler;
    }
 
    /**
@@ -85,57 +64,7 @@ public class VSSortingDialogController {
                                                        @RequestParam("runtimeId") String runtimeId, Principal principal)
       throws Exception
    {
-      RuntimeViewsheet rvs = viewsheetService.getViewsheet(runtimeId, principal);
-      Viewsheet viewsheet = rvs.getViewsheet();
-      CalcTableVSAssembly assembly = (CalcTableVSAssembly) viewsheet.getAssembly(objectId);
-      CalcTableVSAssemblyInfo assemblyInfo = (CalcTableVSAssemblyInfo) assembly.getVSAssemblyInfo();
-      SourceInfo sourceInfo = assemblyInfo.getSourceInfo();
-      SortInfo sortInfo = assemblyInfo.getSortInfo();
-      ColumnSelection columnSelection = null;
-      SortRef[] sortRefs = sortInfo == null ? new SortRef[0] : sortInfo.getSorts();
-      List<VSSortRefModel> columnSortList = new ArrayList<>();
-      List<VSSortRefModel> columnNoneList = new ArrayList<>();
-      DataRef[] allColumns = new DataRef[0];
-
-      if(sourceInfo != null && sourceInfo.getSource() != null) {
-         columnSelection =
-            vsInputService.getTableColumns(rvs, sourceInfo.getSource(), null, null, null,
-                                           false, false, true, false, false, true, principal);
-         allColumns = (DataRef[]) Collections.list(columnSelection.getAttributes()).toArray(new DataRef[0]);
-      }
-
-      for(SortRef sortRef : sortRefs) {
-         if(columnSelection == null || !containsColumn(sortRef.getDataRef(), allColumns)) {
-            continue;
-         }
-
-         VSSortRefModel sortRefModel = new VSSortRefModel();
-         sortRefModel.setDataRefModel(dataRefModelFactoryService.createDataRefModel(sortRef.getDataRef()));
-         sortRefModel.setOrder(sortRef.getOrder());
-         columnSortList.add(sortRefModel);
-      }
-
-      if(columnSelection != null) {
-         for(int i = 0; i < columnSelection.getAttributeCount(); i++) {
-            DataRef ref = columnSelection.getAttribute(i);
-
-            if(containsColumn(ref, sortRefs)) {
-               continue;
-            }
-
-            VSSortRefModel sortRefModel = new VSSortRefModel();
-            sortRefModel.setDataRefModel(dataRefModelFactoryService.createDataRefModel(ref));
-            sortRefModel.setOrder(StyleConstants.SORT_NONE);
-            columnNoneList.add(sortRefModel);
-         }
-      }
-
-      VSSortingDialogModel vsSortingDialogModel = new VSSortingDialogModel();
-      VSSortingPaneModel vsSortingPaneModel = vsSortingDialogModel.getVsSortingPaneModel();
-      vsSortingPaneModel.setColumnSortList(columnSortList.toArray(new VSSortRefModel[0]));
-      vsSortingPaneModel.setColumnNoneList(columnNoneList.toArray(new VSSortRefModel[0]));
-
-      return vsSortingDialogModel;
+      return vsInputServiceProxy.getVSSortingDialogModel(runtimeId, objectId, principal);
    }
 
    @Undoable
@@ -143,45 +72,12 @@ public class VSSortingDialogController {
    public void setVSSortingDialogModel(@DestinationVariable("objectId") String objectId,
                                        @Payload VSSortingDialogModel model,
                                        Principal principal,
-                                       CommandDispatcher dispatcher)
-      throws Exception {
-      RuntimeViewsheet rvs =
-         service.getRuntimeViewsheet(runtimeViewsheetRef.getRuntimeId(), principal);
-      Viewsheet viewsheet = rvs.getViewsheet();
-      CalcTableVSAssembly assembly = (CalcTableVSAssembly) viewsheet.getAssembly(objectId);
-      CalcTableVSAssemblyInfo assemblyInfo = (CalcTableVSAssemblyInfo) assembly.getVSAssemblyInfo();
-      SortInfo sortInfo = new SortInfo();
-
-      for(VSSortRefModel vsSortRefModel : model.getVsSortingPaneModel().getColumnSortList()) {
-         SortRef sortRef = new SortRef();
-         DataRefModel dataRefModel = vsSortRefModel.getDataRefModel();
-         sortRef.setDataRef(dataRefModel.createDataRef());
-         sortRef.setOrder(vsSortRefModel.getOrder());
-         sortInfo.addSort(sortRef);
-      }
-
-      assemblyInfo.setSortInfo(sortInfo);
-      this.vsAssemblyInfoHandler.apply(rvs, assemblyInfo, viewsheetService, false, false, true, false, dispatcher);
+                                       CommandDispatcher dispatcher) throws Exception
+   {
+      vsInputServiceProxy.setVSSortingDialogModel(runtimeViewsheetRef.getRuntimeId(), objectId,
+                                                  model, principal, dispatcher);
    }
 
-   private boolean containsColumn(DataRef ref, DataRef[] refArray) {
-      for(DataRef ref0 : refArray) {
-         if(ref0 instanceof SortRef) {
-            ref0 = ((SortRef) ref0).getDataRef();
-         }
-
-         if(Tool.equals(ref, ref0)) {
-            return true;
-         }
-      }
-
-      return false;
-   }
-
-   private final VSObjectService service;
-   private final VSInputService vsInputService;
-   private final DataRefModelFactoryService dataRefModelFactoryService;
-   private final ViewsheetService viewsheetService;
+   private final VSInputServiceProxy vsInputServiceProxy;
    private final RuntimeViewsheetRef runtimeViewsheetRef;
-   private VSAssemblyInfoHandler vsAssemblyInfoHandler;
 }
