+++ core/src/main/java/inetsoft/web/composer/vs/objects/controller/VSSliderController.java
@@ -20,6 +20,7 @@ package inetsoft.web.composer.vs.objects.controller;
 import inetsoft.web.composer.vs.objects.event.ChangeVSObjectValueEvent;
 import inetsoft.web.viewsheet.LoadingMask;
 import inetsoft.web.viewsheet.Undoable;
+import inetsoft.web.viewsheet.model.RuntimeViewsheetRef;
 import inetsoft.web.viewsheet.service.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -35,14 +36,12 @@ import java.security.Principal;
  */
 @Controller
 public class VSSliderController {
-   /**
-    * Creates a new instance of <tt>VSSliderController</tt>.
-    *
-    * @param vsInputService   reference to the VSInputService
-    */
+
    @Autowired
-   public VSSliderController(VSInputService vsInputService) {
-      this.vsInputService = vsInputService;
+   public VSSliderController(VSInputServiceProxy vsInputServiceProxy,
+                             RuntimeViewsheetRef runtimeViewsheetRef) {
+      this.vsInputServiceProxy = vsInputServiceProxy;
+      this.runtimeViewsheetRef = runtimeViewsheetRef;
    }
 
    /**
@@ -62,11 +61,13 @@ public class VSSliderController {
                            CommandDispatcher dispatcher, @LinkUri String linkUri)
       throws Exception
    {
-      this.vsInputService.singleApplySelection(event.getName(), event.getValue(),
-                                         principal, dispatcher, linkUri);
+      this.vsInputServiceProxy.singleApplySelection(runtimeViewsheetRef.getRuntimeId(),
+                                                    event.getName(), event.getValue(),
+                                                    principal, dispatcher, linkUri);
    }
 
-   private final VSInputService vsInputService;
+   private final RuntimeViewsheetRef runtimeViewsheetRef;
+   private VSInputServiceProxy vsInputServiceProxy;
 
    private static final Logger LOG =
       LoggerFactory.getLogger(VSSliderController.class);
