+++ core/src/main/java/inetsoft/web/viewsheet/controller/VSComboBoxController.java
@@ -17,15 +17,8 @@
  */
 package inetsoft.web.viewsheet.controller;
 
-
-import inetsoft.report.composition.RuntimeViewsheet;
-import inetsoft.report.internal.table.TableFormat;
-import inetsoft.uql.viewsheet.ComboBoxVSAssembly;
-import inetsoft.uql.viewsheet.Viewsheet;
-import inetsoft.uql.viewsheet.internal.ComboBoxVSAssemblyInfo;
 import inetsoft.web.viewsheet.LoadingMask;
 import inetsoft.web.viewsheet.Undoable;
-import inetsoft.web.viewsheet.command.MessageCommand;
 import inetsoft.web.viewsheet.event.VSListInputSelectionEvent;
 import inetsoft.web.viewsheet.model.RuntimeViewsheetRef;
 import inetsoft.web.viewsheet.service.*;
@@ -33,10 +26,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.stereotype.Controller;
-
 import java.security.Principal;
-import java.text.Format;
-import java.util.Date;
+
 
 /**
  * Controller that provides REST endpoints and message handling for Combo Box.
@@ -49,11 +40,10 @@ public class VSComboBoxController {
     * Creates a new instance of <tt>VSComboBoxController</tt>.
     */
    @Autowired
-   public VSComboBoxController(VSObjectService service,
-                               VSInputService inputService, RuntimeViewsheetRef runtimeViewsheetRef)
+   public VSComboBoxController(VSInputServiceProxy inputServiceProxy,
+                               RuntimeViewsheetRef runtimeViewsheetRef)
    {
-      this.service = service;
-      this.inputService = inputService;
+      this.inputServiceProxy = inputServiceProxy;
       this.runtimeViewsheetRef = runtimeViewsheetRef;
    }
 
@@ -74,45 +64,10 @@ public class VSComboBoxController {
                               @LinkUri String linkUri)
       throws Exception
    {
-      final String assemblyName = event.assemblyName();
-      final RuntimeViewsheet rvs =
-         service.getRuntimeViewsheet(runtimeViewsheetRef.getRuntimeId(), principal);
-      final Viewsheet viewsheet = rvs.getViewsheet();
-      final ComboBoxVSAssembly assembly = (ComboBoxVSAssembly) viewsheet.getAssembly(assemblyName);
-      final ComboBoxVSAssemblyInfo info = (ComboBoxVSAssemblyInfo) assembly.getVSAssemblyInfo();
-      Object selectedValue = event.value();
-
-      try {
-         if(info.isCalendar() && selectedValue != null && !"".equals(selectedValue)) {
-            try {
-               selectedValue = new Date(Long.parseLong(selectedValue.toString()));
-            }
-            catch(Exception ex) {
-               String fmt = info.getFormat().getFormat();
-               String spec = info.getFormat().getFormatExtent();
-               Format format = TableFormat.getFormat(fmt, spec);
-
-               if(format != null) {
-                  selectedValue = format.parseObject(selectedValue.toString());
-               }
-               else {
-                  throw ex;
-               }
-            }
-         }
-      }
-      catch(Exception e) {
-         MessageCommand command = new MessageCommand();
-         command.setMessage(e.getMessage());
-         command.setType(MessageCommand.Type.ERROR);
-         dispatcher.sendCommand(command);
-         return;
-      }
-
-      inputService.singleApplySelection(assemblyName, selectedValue, principal, dispatcher, linkUri);
+      inputServiceProxy.applySelection(runtimeViewsheetRef.getRuntimeId(), event,
+                                       principal, dispatcher, linkUri);
    }
 
-   private final VSObjectService service;
-   private final VSInputService inputService;
+   private final VSInputServiceProxy inputServiceProxy;
    private final RuntimeViewsheetRef runtimeViewsheetRef;
 }
